+++ src/User/Controller/RecoveryController.php
@@ -63,7 +63,7 @@ class RecoveryController extends Controller
     {
         return [
             'access' => [
-                'class' => AccessControl::className(),
+                'class' => AccessControl::class,
                 'rules' => [
                     [
                         'allow' => true,