+++ src/User/Controller/RegistrationController.php
@@ -72,7 +72,7 @@ class RegistrationController extends Controller
     {
         return [
             'access' => [
-                'class' => AccessControl::className(),
+                'class' => AccessControl::class,
                 'rules' => [
                     [
                         'allow' => true,