+++ src/User/resources/views/role/index.php
@@ -54,7 +54,7 @@ $this->params['breadcrumbs'][] = $this->title;
                 ],
             ],
             [
-                'class' => ActionColumn::className(),
+                'class' => ActionColumn::class,
                 'template' => '{update} {delete}',
                 'urlCreator' => function ($action, $model) {
                     return Url::to(['/user/role/' . $action, 'name' => $model['name']]);