+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  - Fix #195: UserCreateService: check if we're from web before setting flash message (maxxer)
  - Enh: Improvements to the admin responsive design (wautvda)
  - Enh: Add controller module class reference (TonisOrmisson)
+ - Enh: Replace the deprecated InvalidParamException in ClassMapHelper (TonisOrmisson)
  
 ## 1.1.4 - February 19, 2018
 - Enh: Check enableEmailConfirmation on registration (faenir)